入力エリア部品化
